move to ios and format
