--- dataline-commons/src/main/java/io/dataline/commons/io/TailHelper.java
@@ -1,58 +1,0 @@
-/*
- * MIT License
- *
- * Copyright (c) 2020 Dataline
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all
- * copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- */
-
-package io.dataline.commons.io;
-
-import com.google.common.base.Charsets;
-import java.io.File;
-import java.io.IOException;
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.List;
-import org.apache.commons.io.input.ReversedLinesFileReader;
-
-public class TailHelper {
-
-  public static List<String> getTail(int numLines, String path) throws IOException {
-    File file = new File(path);
-
-    if (file.exists()) {
-      try (ReversedLinesFileReader fileReader = new ReversedLinesFileReader(file, Charsets.UTF_8)) {
-        List<String> lines = new ArrayList<>();
-
-        String line;
-        while ((line = fileReader.readLine()) != null && lines.size() < numLines) {
-          lines.add(line);
-        }
-
-        Collections.reverse(lines);
-
-        return lines;
-      }
-    } else {
-      return Collections.emptyList();
-    }
-  }
-
-}
